alexa-voiceagent: fixed the imported libraries
Since with a recent AGL SDK, the symlinks to files
are copied in the package directory, in place of
the actual pointed files, and that the widget format
does not (yet) support symlinks, at the end the imported
libraries that are pointed by symlinks were not part
of the widget.
The fix here consists in making the copy directly,
replacing the link (but keeping its name) by the
pointed file.

Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>